+++ lib/Api/LoyaltyCardgroupsApi.php
@@ -263,7 +263,7 @@ class LoyaltyCardgroupsApi
     /**
      * Operation getAll
      *
-     * Find Loyalty Cardgroups
+     * Find Loyalty Card Groups
      *
      * @param int $count The maximum number of items to return 
      * @param int $offset The position within the whole result set to start returning items (zero-based) 
@@ -288,7 +288,7 @@ class LoyaltyCardgroupsApi
     /**
      * Operation getAllWithHttpInfo
      *
-     * Find Loyalty Cardgroups
+     * Find Loyalty Card Groups
      *
      * @param int $count The maximum number of items to return 
      * @param int $offset The position within the whole result set to start returning items (zero-based) 
@@ -393,107 +393,6 @@ class LoyaltyCardgroupsApi
                 switch ($e->getCode()) {
                     case 200:
                         $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\LoyaltyCardgroupsList', $e->getResponseHeaders());
-                        $e->setResponseObject($data);
-                        break;
-                    case 401:
-                        if ($retries < 1) {
-                            Authenticator::reauthenticate();
-                            continue 2;
-                        }
-                    default:
-                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
-                        $e->setResponseObject($data);
-                        break;
-                }
-
-                throw $e;
-            }
-        }
-    }
-
-    /**
-     * Operation getCardgroupsByMerchantId
-     *
-     * Find Loyalty Sales by General Merchant ID
-     *
-     * @param string $general_merchant_id Search by provided id (required)
-     * @throws ApiException on non-2xx response
-     * @return \Secuconnect\Client\Model\LoyaltyCardgroupsProductModel[]
-     */
-    public function getCardgroupsByMerchantId($general_merchant_id)
-    {
-        list($response) = $this->getCardgroupsByMerchantIdWithHttpInfo($general_merchant_id);
-        return $response;
-    }
-
-    /**
-     * Operation getCardgroupsByMerchantIdWithHttpInfo
-     *
-     * Find Loyalty Sales by General Merchant ID
-     *
-     * @param string $general_merchant_id Search by provided id (required)
-     * @throws ApiException on non-2xx response
-     * @return array of \Secuconnect\Client\Model\LoyaltyCardgroupsProductModel[], HTTP status code, HTTP response headers (array of strings)
-     */
-    public function getCardgroupsByMerchantIdWithHttpInfo($general_merchant_id)
-    {
-        // verify the required parameter 'general_merchant_id' is set
-        if ($general_merchant_id === null || (is_array($general_merchant_id) && count($general_merchant_id) === 0)) {
-            throw new \InvalidArgumentException(
-                'Missing the required parameter $general_merchant_id when calling getCardgroupsByMerchantId'
-            );
-        }
-        // parse inputs
-        $resourcePath = "/Loyalty/Sales/{generalMerchantId}/CardGroupsByMerchantID";
-        $httpBody = '';
-        $queryParams = [];
-        $headerParams = [];
-        $formParams = [];
-        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
-        if (!is_null($_header_accept)) {
-            $headerParams['Accept'] = $_header_accept;
-        }
-        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
-
-        // path params
-        if ($general_merchant_id !== null) {
-            $resourcePath = str_replace(
-                "{" . "generalMerchantId" . "}",
-                $this->apiClient->getSerializer()->toPathValue($general_merchant_id),
-                $resourcePath
-            );
-        }
-
-        // for model (json/xml)
-        if (isset($_tempBody)) {
-            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
-        } elseif (count($formParams) > 0) {
-            $httpBody = $formParams; // for HTTP post (form)
-        }
-        for ($retries = 0; ; $retries++) {
-
-            // this endpoint requires OAuth (access token)
-            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
-                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
-            }
-
-            // make the API Call
-            try {
-                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                    $resourcePath,
-                    'GET',
-                    $queryParams,
-                    $httpBody,
-                    $headerParams,
-                    '\Secuconnect\Client\Model\LoyaltyCardgroupsProductModel[]',
-                    '/Loyalty/Sales/{generalMerchantId}/CardGroupsByMerchantID'
-                );
-
-                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\LoyaltyCardgroupsProductModel[]', $httpHeader), $statusCode, $httpHeader];
-            } catch (ApiException $e) {
-                switch ($e->getCode()) {
-                    case 200:
-                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\LoyaltyCardgroupsProductModel[]', $e->getResponseHeaders());
                         $e->setResponseObject($data);
                         break;
                     case 401:
@@ -619,7 +518,7 @@ class LoyaltyCardgroupsApi
      * Update Loyalty Cardgroup
      *
      * @param string $loyalty_card_group_id Loyalty card group id (required)
-     * @param \Secuconnect\Client\Model\LoyaltyCardgroupsDTO $body Loyalty card group properties 
+     * @param \Secuconnect\Client\Model\LoyaltyCardgroupsUpdateDTO $body Loyalty card group properties 
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\LoyaltyCardgroupsProductModel
      */
@@ -635,7 +534,7 @@ class LoyaltyCardgroupsApi
      * Update Loyalty Cardgroup
      *
      * @param string $loyalty_card_group_id Loyalty card group id (required)
-     * @param \Secuconnect\Client\Model\LoyaltyCardgroupsDTO $body Loyalty card group properties 
+     * @param \Secuconnect\Client\Model\LoyaltyCardgroupsUpdateDTO $body Loyalty card group properties 
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\LoyaltyCardgroupsProductModel, HTTP status code, HTTP response headers (array of strings)
      */